// resources/js/enums/EventLocationTypes.tsx
enum EventLocationTypes {
    TEXT = 'text',
    IN_PERSON_HOST = 'in_person_host',
    IN_PERSON_GUEST = 'in_person_guest',
    CUSTOM_LINK = 'link',
    PHONE_INCOMING = 'phone_incoming',
    PHONE_OUTGOING = 'phone_outgoing',
    GOOGLE_MEET = 'google_meet',
    ZOOM = 'zoom',
    MICROSOFT_TEAMS = 'microsoft_teams',
}

export default EventLocationTypes;